refactor:将ai相关的代码与服务层代码分离
