+++ src/main/java/com/example/aitourism/ai/AssistantService.java
@@ -1,10 +1,10 @@
-package com.example.aitourism.service;
+package com.example.aitourism.ai;
 
 import dev.langchain4j.service.MemoryId;
 import dev.langchain4j.service.TokenStream;
 
 // 定义Assistant接口
-public interface Assistant{
+public interface AssistantService {
     @dev.langchain4j.service.SystemMessage(fromResource ="prompt/tour-route-planning-system-prompt.txt")
     TokenStream chat_Stream(@MemoryId String memoryId, @dev.langchain4j.service.UserMessage String userMessage);
